labs/DB/lab5/code/requirements.txt
PyQt6>=6.4.0
psycopg2-binary>=2.9.0
